+++ src/components/settings/ExportLibrary.js
@@ -5,7 +5,7 @@ import Button from 'react-bootstrap/Button';
 
 const ExportLibrary = (props) => {
     return (
-        <Row className="p-0 ml-0 mr-0 tm-40">
+        <Row className="pl-5 pr-5 pt-5 m-0">
             <Col xs={12} md={10} className="p-0 m-0">
                 <h5 className="bold-text">Export Library</h5>
                 <p className="text-light m-0">Export your music library as a JSON file. It can then be shared with other Hope Player users.</p>
