+++ src/pages/settings.js
@@ -9,7 +9,7 @@ import ImportLibrary from '../components/settings/ImportLibrary';
 const Settings = () => {
     return (
         <div className="page-root settings-container">
-            <h3 className="page-headline mb-0 lm-40 tm-40 rm-40 bottom-border">Settings</h3>
+            <h3 className="page-headline mt-5 ml-5 mr-5 mb-0 bottom-border">Settings</h3>
             <div className="settings-list-container">
                 <AudioQuality />
